consistent ergodic cut off
