+++ msmbuilder/tests/test_msm.py
@@ -15,7 +15,7 @@ from msmbuilder.utils import map_drawn_samples
 from msmbuilder import cluster
 
 from msmbuilder.msm.core import _transition_counts
-from msmbuilder.msm import MarkovStateModel
+from msmbuilder.msm import MarkovStateModel, BayesianMarkovStateModel
 
 
 def test_counts_1():
@@ -356,3 +356,9 @@ def test_score_1():
 
         assert_approx_equal(model.score([sequence]), model.eigenvalues_.sum())
         assert_approx_equal(model.score([sequence]), model.score_)
+
+def test_ergodic_cut_off():
+    for cut_off in [0.01, 'on', 'off']:
+        assert(MarkovStateModel(ergodic_cutoff=cut_off).ergodic_cutoff ==
+               BayesianMarkovStateModel(ergodic_cutoff=cut_off).ergodic_cutoff)
+    return
